+++ tests/integration/test_payment_method_nonce.py
@@ -103,15 +103,15 @@ class TestPaymentMethodNonce(unittest.TestCase):
 
         self.assertEqual("CreditCard", found_nonce.type)
         self.assertEqual(nonce, found_nonce.nonce)
-        self.assertEqual("Y", three_d_secure_info.enrolled)
         self.assertEqual("authenticate_successful", three_d_secure_info.status)
         self.assertEqual(True, three_d_secure_info.liability_shifted)
         self.assertEqual(True, three_d_secure_info.liability_shift_possible)
-        self.assertEqual("cavv_value", three_d_secure_info.cavv)
-        self.assertEqual("xid_value", three_d_secure_info.xid)
-        self.assertEqual("05", three_d_secure_info.eci_flag)
-        self.assertEqual("1.0.2", three_d_secure_info.three_d_secure_version)
-        self.assertIsNotNone(three_d_secure_info.three_d_secure_authentication_id)
+        self.assertIsInstance(three_d_secure_info.enrolled, str)
+        self.assertIsInstance(three_d_secure_info.cavv, str)
+        self.assertIsInstance(three_d_secure_info.xid, str)
+        self.assertIsInstance(three_d_secure_info.eci_flag, str)
+        self.assertIsInstance(three_d_secure_info.three_d_secure_version, str)
+        self.assertIsInstance(three_d_secure_info.three_d_secure_authentication_id, str)
 
     def test_find_nonce_shows_paypal_details(self):
         found_nonce = PaymentMethodNonce.find("fake-google-pay-paypal-nonce")